requirements.txt
pyTelegramBotAPI==4.0.1
huggingface-hub==0.21.1
transformers==4.39.0
python-telegram-bot==20.3
openai==0.27.8

